refactor(runtime): split vm_arith_ops.rs into cohesive modules (§7-8)
`vm_arith_ops.rs` was 1639 lines, over the 500-line module limit. Split it into
five files along its natural seams, each under 500 lines, with no behavior
change (the set of fn signatures is identical before/after):

- `vm_arith_ops.rs` (250) — core arithmetic ops: add/sub/mul/div/mod/pow,
  negate, `try_user_infix`, and the int/bool/str bit-negation ops.
- `vm_coerce_concat_ops.rs` (322) — decont/slip, `not`/bool coercion, string
  concatenation (with junction threading), and Buf-value helpers.
- `vm_arith_int_ops.rs` (490) — integer division/mod, gcd/lcm, infix min/max,
  string/list repeat (incl. the `xx` reeval-thunk helpers moved here from their
  sole caller `exec_list_repeat_op`), and function composition.
- `vm_mixin_does_ops.rs` (348) — `but`/`does` mixin ops, `isa`, and pair
  construction, plus their private mixin/does helpers.
- `vm_bitwise_ops.rs` (253) — bitwise and/or/xor and shift ops for int, bool,
  and string operands.

The three `xx`-thunk helpers (`is_xx_reeval_thunk`, `try_bulk_shift_pop`,
`vm_xx_repeat_thunk`) move with `exec_list_repeat_op` (their only caller), so no
visibility change is needed; all other private helpers stay with their callers.
Each file carries only the imports it uses. Seventh slice of the §7-8
file-splitting sweep.
