+++ src/vm/vm_arith_ops.rs
@@ -1,189 +1,6 @@
 use super::*;
-use num_traits::ToPrimitive;
-use std::sync::Arc;
-use unicode_normalization::UnicodeNormalization;
 
 impl Interpreter {
-    fn is_xx_reeval_thunk(data: &crate::value::SubData) -> bool {
-        if !data.params.is_empty()
-            || !data.param_defs.is_empty()
-            || !data.assumed_positional.is_empty()
-            || !data.assumed_named.is_empty()
-            || data.body.len() != 1
-        {
-            return false;
-        }
-        match &data.body[0] {
-            crate::ast::Stmt::Expr(expr) => matches!(
-                expr,
-                crate::ast::Expr::Call { .. }
-                    | crate::ast::Expr::MethodCall { .. }
-                    | crate::ast::Expr::DynamicMethodCall { .. }
-                    | crate::ast::Expr::HyperMethodCall { .. }
-                    | crate::ast::Expr::HyperMethodCallDynamic { .. }
-                    | crate::ast::Expr::CallOn { .. }
-                    | crate::ast::Expr::Block(_)
-                    | crate::ast::Expr::DoBlock { .. }
-                    | crate::ast::Expr::AnonSub { .. }
-                    | crate::ast::Expr::AnonSubParams { .. }
-                    | crate::ast::Expr::Lambda { .. }
-            ),
-            _ => false,
-        }
-    }
-
-    /// Fast path for `@var.shift xx N` and `@var.pop xx N`.
-    /// Drains N elements in bulk instead of running a thunk N times.
-    fn try_bulk_shift_pop(
-        &mut self,
-        data: &crate::value::SubData,
-        n: usize,
-    ) -> Result<Option<Vec<Value>>, RuntimeError> {
-        use crate::ast::{Expr, Stmt};
-        if data.body.len() != 1 {
-            return Ok(None);
-        }
-        let (var_name, is_shift) = match &data.body[0] {
-            Stmt::Expr(Expr::MethodCall {
-                target, name, args, ..
-            }) if args.is_empty() && (name.resolve() == "shift" || name.resolve() == "pop") => {
-                let vname = match target.as_ref() {
-                    Expr::Var(v) => v.clone(),
-                    Expr::ArrayVar(v) => format!("@{}", v),
-                    _ => return Ok(None),
-                };
-                (vname, name.resolve() == "shift")
-            }
-            _ => return Ok(None),
-        };
-        // Look up the array in env
-        let arr_key = if var_name.starts_with('@') {
-            var_name.clone()
-        } else {
-            format!("@{}", var_name)
-        };
-        let target_val = self
-            .env()
-            .get(&arr_key)
-            .or_else(|| self.env().get(&var_name))
-            .cloned();
-        let Value::Array(arc_items, kind) = target_val.unwrap_or(Value::Nil) else {
-            return Ok(None);
-        };
-        if !kind.is_real_array() {
-            return Ok(None);
-        }
-        let mut items = (*arc_items).clone();
-        let actual_n = n.min(items.len());
-        let result = if is_shift {
-            let rest = items.split_off(actual_n);
-            let shifted = items;
-            items = crate::value::ArrayData::new(rest);
-            shifted
-        } else {
-            // pop
-            let split_at = items.len().saturating_sub(actual_n);
-            let popped: Vec<Value> = items.drain(split_at..).rev().collect();
-            crate::value::ArrayData::new(popped)
-        };
-        // Write the mutated array back
-        let lookup_key = if self.env().contains_key(&arr_key) {
-            &arr_key
-        } else {
-            &var_name
-        };
-        self.env_mut()
-            .insert(lookup_key.to_string(), Value::Array(Arc::new(items), kind));
-        Ok(Some(result.items))
-    }
-
-    /// Interpreter-native implementation of xx-repeat for thunks.
-    /// Compiles the thunk body once and runs it N times via `run_reuse`,
-    /// avoiding the interpreter roundtrip through `eval_xx_repeat_thunk`.
-    fn vm_xx_repeat_thunk(
-        &mut self,
-        data: &crate::value::SubData,
-        n: usize,
-    ) -> Result<Vec<Value>, RuntimeError> {
-        // Save env keys that the thunk captures
-        let touched_keys: Vec<String> = data.env.keys().map(|k| k.resolve()).collect();
-        let saved: Vec<(String, Option<Value>)> = touched_keys
-            .iter()
-            .map(|k| (k.clone(), self.env().get(k).cloned()))
-            .collect();
-
-        // Install captured env
-        for (k, v) in data.env.iter() {
-            if matches!(self.env().get_sym(*k), Some(Value::Array(..)))
-                && matches!(v, Value::Array(..))
-            {
-                continue;
-            }
-            self.env_mut().insert_sym(*k, v.clone());
-        }
-
-        // Compile the thunk body
-        let compiler = crate::compiler::Compiler::new();
-        let (code, compiled_fns) = compiler.compile(&data.body);
-        let code = Arc::new(code);
-
-        // Run N times, collecting results
-        let mut out = Vec::with_capacity(n);
-        let saved_stack = std::mem::take(&mut self.stack);
-        let saved_locals = std::mem::take(&mut self.locals);
-
-        for _ in 0..n {
-            match self.run_reuse(&code, &compiled_fns) {
-                Ok(()) => {
-                    let val = self.env().get("_").cloned().unwrap_or(Value::Nil);
-                    out.push(val);
-                }
-                Err(e) => {
-                    self.stack = saved_stack;
-                    self.locals = saved_locals;
-                    // Restore env
-                    for (k, orig) in saved {
-                        match orig {
-                            Some(v) => {
-                                self.env_mut().insert(k, v);
-                            }
-                            None => {
-                                self.env_mut().remove(&k);
-                            }
-                        }
-                    }
-                    return Err(e);
-                }
-            }
-        }
-
-        self.stack = saved_stack;
-        self.locals = saved_locals;
-        // Restore env, but keep mutations to arrays (e.g. shift/pop
-        // inside the thunk should persist).
-        for (k, orig) in saved {
-            // If the thunk mutated an array variable (e.g. via .shift),
-            // keep the mutated version instead of restoring the original.
-            let current = self.env().get(&k).cloned();
-            let should_keep_current = matches!(
-                (&orig, &current),
-                (Some(Value::Array(..)), Some(Value::Array(..)))
-            );
-            if should_keep_current {
-                continue;
-            }
-            match orig {
-                Some(v) => {
-                    self.env_mut().insert(k, v);
-                }
-                None => {
-                    self.env_mut().remove(&k);
-                }
-            }
-        }
-        Ok(out)
-    }
-
     pub(super) fn exec_add_op(&mut self) -> Result<(), RuntimeError> {
         let right = self.stack.pop().unwrap();
         let left = self.stack.pop().unwrap();
@@ -428,1212 +245,5 @@ impl Interpreter {
             let result = String::from_utf8_lossy(&negated).into_owned();
             self.stack.push(Value::str(result));
         }
-    }
-
-    pub(super) fn exec_decont_op(&mut self) {
-        // Strips a SINGLE level of Value::Scalar for slurpy flattening. This is
-        // intentionally non-recursive and distinct from the recursive
-        // Value::descalarize (see the decont family note in value/mod.rs §3).
-        let val = self.stack.pop().unwrap();
-        let new_val = match val {
-            Value::Scalar(inner) => *inner,
-            other => other,
-        };
-        self.stack.push(new_val);
-    }
-
-    /// Pull every element from a deferred-iterator `Seq` (as stored by
-    /// `Seq.new($iterator)`), driving the iterator's `pull-one` until
-    /// `IterationEnd`. Works for both built-in and user-defined `Iterator`
-    /// classes. The deferred iterator is removed and the Seq is marked consumed.
-    pub(super) fn materialize_deferred_seq(&mut self, items_arc: &Arc<Vec<Value>>) -> Vec<Value> {
-        let Some(iterator) = crate::value::seq_take_deferred_iter(items_arc) else {
-            return (**items_arc).clone();
-        };
-        crate::value::seq_consume(items_arc).ok();
-        let mut pulled = Vec::new();
-        while let Ok(val) = self.call_method_with_values(iterator.clone(), "pull-one", vec![]) {
-            if matches!(&val, Value::Str(s) if s.as_str() == "IterationEnd")
-                || matches!(&val, Value::Package(name) if *name == crate::symbol::Symbol::intern("IterationEnd"))
-            {
-                break;
-            }
-            pulled.push(val);
-        }
-        pulled
-    }
-
-    pub(super) fn exec_make_slip_op(&mut self) {
-        let val = self.stack.pop().unwrap();
-        // Slipping (`|EXPR`) always flattens through containers/itemization, e.g.
-        // `|$_` where the topic is an itemized Seq element must expand the Seq's
-        // values (`($seq,).map(|*)`), not wrap the Seq as a single slip item.
-        let val = val.into_deref().into_descalarized();
-        // A `Seq.new($iterator)` stores its iterator deferred (empty backing vec).
-        // Slipping such a Seq must first pull all elements from the iterator
-        // (including user-defined `Iterator` classes), else `|$seq` yields nothing.
-        if let Value::Seq(items_arc) = &val
-            && crate::value::seq_has_deferred_iter(items_arc)
-        {
-            let pulled = self.materialize_deferred_seq(items_arc);
-            self.stack.push(Value::slip(pulled));
-            return;
-        }
-        let items = match &val {
-            Value::Array(items, ..) => (*items).to_vec(),
-            Value::Slip(items) => (*items).to_vec(),
-            Value::Seq(items) => (*items).to_vec(),
-            Value::Capture { positional, named } => {
-                let mut items = (**positional).clone();
-                for (k, v) in named.iter() {
-                    items.push(Value::Pair(k.clone(), Box::new(v.clone())));
-                }
-                items
-            }
-            Value::Hash(map) => map
-                .iter()
-                .map(|(k, v)| Value::Pair(k.clone(), Box::new(v.clone())))
-                .collect(),
-            Value::LazyList(ll) => {
-                if ll.scan_spec.is_some() {
-                    ll.force_scan_to(200_000)
-                } else {
-                    ll.cache.lock().unwrap().clone().unwrap_or_default()
-                }
-            }
-            Value::LazyIoLines { .. } => match self.force_if_lazy_io_lines(val) {
-                Ok(forced) => crate::runtime::utils::value_to_list(&forced),
-                Err(_) => vec![],
-            },
-            Value::Range(..)
-            | Value::RangeExcl(..)
-            | Value::RangeExclStart(..)
-            | Value::RangeExclBoth(..)
-            | Value::GenericRange { .. } => crate::runtime::utils::value_to_list(&val),
-            _ => vec![val],
-        };
-        self.stack.push(Value::slip(items));
-    }
-
-    pub(super) fn exec_not_op(&mut self) {
-        let val = self.stack.pop().unwrap();
-        // Boolifying a Failure marks it as handled
-        val.mark_failure_handled();
-        let t = self.eval_truthy(&val);
-        self.stack.push(Value::Bool(!t));
-    }
-
-    pub(super) fn exec_bool_coerce_op(&mut self) {
-        let val = self.stack.pop().unwrap();
-        let out = match &val {
-            Value::Regex(_)
-            | Value::RegexWithAdverbs { .. }
-            | Value::Routine { is_regex: true, .. } => {
-                let topic = self.env().get("_").cloned().unwrap_or(Value::Nil);
-                Value::Bool(self.vm_smart_match(&topic, &val))
-            }
-            _ => {
-                // Boolifying a Failure marks it as handled
-                val.mark_failure_handled();
-                Value::Bool(self.eval_truthy(&val))
-            }
-        };
-        self.stack.push(out);
-    }
-
-    pub(super) fn exec_concat_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        // Thread over junctions — concat uses left-first threading
-        // (unlike arithmetic/comparison which uses right-first for tighter
-        // junctions). When both operands are junctions and the right is
-        // tighter, we thread left first and swap the junction kinds.
-        if matches!(left, Value::Junction { .. }) || matches!(right, Value::Junction { .. }) {
-            let result = self.eval_concat_with_junctions(left, right);
-            self.stack.push(result);
-            return Ok(());
-        }
-        // Infix `~` stringifies an operand via `.Stringy` (falling back to `.Str`),
-        // so an operand whose class defines a user `Stringy`/`Str` must dispatch it
-        // — the pure `concat_values` only knows `.gist` (rendering `Foo()`). This is
-        // an internal redispatch with no surrounding CallMethod op, so drain any
-        // captured-outer writeback into the caller's slot (Slice 1b render pattern).
-        let caller_code = self.current_code;
-        let left = self.concat_coerce_operand(left)?;
-        let right = self.concat_coerce_operand(right)?;
-        self.reconcile_caller_after_internal_dispatch(caller_code);
-        let result = Self::concat_values(left, right);
-        self.stack.push(result);
-        Ok(())
-    }
-
-    /// Coerce a concat operand whose class defines a user `Stringy`/`Str` to its
-    /// string value (Raku infix `~` uses `.Stringy`, falling back to `.Str`). Plain
-    /// values and instances without a user stringifier pass through unchanged (the
-    /// pure `concat_values` handles those, including built-in `.gist`/`.Str`).
-    fn concat_coerce_operand(&mut self, v: Value) -> Result<Value, RuntimeError> {
-        let cn = match &v {
-            Value::Instance { class_name, .. } => class_name.resolve().to_string(),
-            Value::Package(name) => name.resolve().to_string(),
-            _ => return Ok(v),
-        };
-        if self.has_user_method(&cn, "Stringy") {
-            let r = self.try_compiled_method_or_interpret(v, "Stringy", Vec::new())?;
-            return Ok(Value::str(r.to_string_value()));
-        }
-        if self.has_user_method(&cn, "Str") {
-            let r = self.try_compiled_method_or_interpret(v, "Str", Vec::new())?;
-            return Ok(Value::str(r.to_string_value()));
-        }
-        Ok(v)
-    }
-
-    fn eval_concat_with_junctions(&mut self, left: Value, right: Value) -> Value {
-        // Auto-FETCH and decontainerize
-        let left = self
-            .auto_fetch_proxy(&left)
-            .unwrap_or(left)
-            .descalarize()
-            .clone();
-        let right = self
-            .auto_fetch_proxy(&right)
-            .unwrap_or(right)
-            .descalarize()
-            .clone();
-        // Both junctions: thread left first, swap kinds if right is tighter
-        if let (Value::Junction { kind: lk, .. }, Value::Junction { kind: rk, .. }) =
-            (&left, &right)
-        {
-            let need_swap = Self::thread_right_first(lk, rk);
-            let rk = rk.clone();
-            let lk = lk.clone();
-            if let Value::Junction { kind, values } = left {
-                let results: Vec<Value> = values
-                    .iter()
-                    .cloned()
-                    .map(|v| self.eval_concat_with_junctions(v, right.clone()))
-                    .collect();
-                let mut result = Value::junction(kind, results);
-                if need_swap {
-                    result = Self::swap_junction_kinds(result, &rk, &lk);
-                }
-                return result;
-            }
-        }
-        if let Value::Junction { kind, values } = left {
-            let results: Vec<Value> = values
-                .iter()
-                .cloned()
-                .map(|v| self.eval_concat_with_junctions(v, right.clone()))
-                .collect();
-            return Value::junction(kind, results);
-        }
-        if let Value::Junction { kind, values } = right {
-            let results: Vec<Value> = values
-                .iter()
-                .cloned()
-                .map(|v| self.eval_concat_with_junctions(left.clone(), v))
-                .collect();
-            return Value::junction(kind, results);
-        }
-        Self::concat_values(left, right)
-    }
-
-    fn swap_junction_kinds(
-        value: Value,
-        new_outer: &crate::value::JunctionKind,
-        new_inner: &crate::value::JunctionKind,
-    ) -> Value {
-        if let Value::Junction { values, .. } = value {
-            let swapped: Vec<Value> = values
-                .iter()
-                .map(|v| {
-                    if let Value::Junction { values: inner, .. } = v {
-                        Value::junction(new_inner.clone(), inner.to_vec())
-                    } else {
-                        v.clone()
-                    }
-                })
-                .collect();
-            Value::junction(new_outer.clone(), swapped)
-        } else {
-            value
-        }
-    }
-
-    /// String/Buf concatenation (`~`). This is the single authoritative impl,
-    /// shared by the Interpreter's `~` op and the interpreter's reduction-operator path
-    /// (`apply_reduction_op` delegates here). It uses no Interpreter state, so it is a
-    /// plain associated function callable as `crate::runtime::Interpreter::concat_values(...)`.
-    pub(crate) fn concat_values(left: Value, right: Value) -> Value {
-        // Buf ~ Buf → Buf (byte concatenation, preserving LHS type)
-        if Self::is_buf_value(&left) && Self::is_buf_value(&right) {
-            let result_class = if let Value::Instance { class_name, .. } = &left {
-                *class_name
-            } else {
-                crate::symbol::Symbol::intern("Buf")
-            };
-            let mut bytes = Self::extract_buf_bytes(&left);
-            bytes.extend(Self::extract_buf_bytes(&right));
-            let byte_vals: Vec<Value> = bytes.into_iter().map(|b| Value::Int(b as i64)).collect();
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("bytes".to_string(), Value::array(byte_vals));
-            return Value::make_instance(result_class, attrs);
-        }
-        // Buf ~ non-Buf or non-Buf ~ Buf: decode the Buf and produce a Str
-        if Self::is_buf_value(&left) || Self::is_buf_value(&right) {
-            let left_str = if Self::is_buf_value(&left) {
-                let bytes = Self::extract_buf_bytes(&left);
-                String::from_utf8_lossy(&bytes).into_owned()
-            } else {
-                crate::runtime::utils::coerce_to_str(&left)
-            };
-            let right_str = if Self::is_buf_value(&right) {
-                let bytes = Self::extract_buf_bytes(&right);
-                String::from_utf8_lossy(&bytes).into_owned()
-            } else {
-                crate::runtime::utils::coerce_to_str(&right)
-            };
-            let concatenated = format!("{}{}", left_str, right_str);
-            if concatenated.is_ascii() {
-                return Value::str(concatenated);
-            }
-            let normalized: String = concatenated.nfc().collect();
-            return Value::str(normalized);
-        }
-        let concatenated = format!(
-            "{}{}",
-            crate::runtime::utils::coerce_to_str(&left),
-            crate::runtime::utils::coerce_to_str(&right)
-        );
-        if concatenated.is_ascii() {
-            Value::str(concatenated)
-        } else {
-            let normalized: String = concatenated.nfc().collect();
-            Value::str(normalized)
-        }
-    }
-
-    pub fn is_buf_value(val: &Value) -> bool {
-        if let Value::Instance { class_name, .. } = val {
-            let cn = class_name.resolve();
-            cn == "Buf"
-                || cn == "Blob"
-                || cn == "utf8"
-                || cn == "utf16"
-                || cn.starts_with("Buf[")
-                || cn.starts_with("Blob[")
-                || cn.starts_with("buf")
-                || cn.starts_with("blob")
-        } else {
-            false
-        }
-    }
-
-    pub fn extract_buf_bytes(val: &Value) -> Vec<u8> {
-        if let Value::Instance { attributes, .. } = val
-            && let Some(Value::Array(items, ..)) = attributes.as_map().get("bytes")
-        {
-            return items
-                .iter()
-                .map(|v| match v {
-                    Value::Int(i) => *i as u8,
-                    _ => 0,
-                })
-                .collect();
-        }
-        Vec::new()
-    }
-
-    pub(super) fn exec_int_div_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let result = self.eval_binary_with_junctions(left, right, |_, l, r| {
-            let val = match (&l, &r) {
-                (Value::Int(a), Value::Int(b)) if *b != 0 => {
-                    Value::Int(num_integer::Integer::div_floor(a, b))
-                }
-                (Value::Int(a), Value::Int(_)) => {
-                    RuntimeError::divide_by_zero_failure(Some(Value::Int(*a)), Some("div"))
-                }
-                (Value::BigInt(a), Value::BigInt(b)) if **b != num_bigint::BigInt::from(0i64) => {
-                    Value::from_bigint(num_integer::Integer::div_floor(a.as_ref(), b.as_ref()))
-                }
-                (Value::BigInt(a), Value::Int(b)) if *b != 0 => {
-                    let bb = num_bigint::BigInt::from(*b);
-                    Value::from_bigint(num_integer::Integer::div_floor(a.as_ref(), &bb))
-                }
-                (Value::Int(a), Value::BigInt(b)) if **b != num_bigint::BigInt::from(0i64) => {
-                    let aa = num_bigint::BigInt::from(*a);
-                    Value::from_bigint(num_integer::Integer::div_floor(&aa, b.as_ref()))
-                }
-                _ => {
-                    let a = runtime::to_int(&l);
-                    let b = runtime::to_int(&r);
-                    if b == 0 {
-                        return Ok(RuntimeError::divide_by_zero_failure(
-                            Some(Value::Int(a)),
-                            Some("div"),
-                        ));
-                    }
-                    Value::Int(num_integer::Integer::div_floor(&a, &b))
-                }
-            };
-            Ok(val)
-        })?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    pub(super) fn exec_int_mod_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let result = self.eval_binary_with_junctions(left, right, |_, l, r| {
-            let val = match (&l, &r) {
-                (Value::Int(a), Value::Int(b)) if *b != 0 => {
-                    Value::Int(num_integer::Integer::mod_floor(a, b))
-                }
-                (Value::Int(a), Value::Int(_)) => {
-                    RuntimeError::divide_by_zero_failure(Some(Value::Int(*a)), Some("%"))
-                }
-                (Value::BigInt(a), Value::BigInt(b)) if !b.is_zero() => {
-                    Value::from_bigint(num_integer::Integer::mod_floor(a.as_ref(), b.as_ref()))
-                }
-                (Value::BigInt(a), Value::Int(b)) if *b != 0 => {
-                    let bb = num_bigint::BigInt::from(*b);
-                    Value::from_bigint(num_integer::Integer::mod_floor(a.as_ref(), &bb))
-                }
-                (Value::Int(a), Value::BigInt(b)) if !b.is_zero() => {
-                    let aa = num_bigint::BigInt::from(*a);
-                    Value::from_bigint(num_integer::Integer::mod_floor(&aa, b.as_ref()))
-                }
-                _ => {
-                    let a = runtime::to_int(&l);
-                    let b = runtime::to_int(&r);
-                    if b == 0 {
-                        return Ok(RuntimeError::divide_by_zero_failure(
-                            Some(Value::Int(a)),
-                            Some("%"),
-                        ));
-                    }
-                    Value::Int(num_integer::Integer::mod_floor(&a, &b))
-                }
-            };
-            Ok(val)
-        })?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    pub(super) fn exec_gcd_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let a = left.to_bigint().abs();
-        let b = right.to_bigint().abs();
-        let g = num_integer::Integer::gcd(&a, &b);
-        self.stack.push(Value::from_bigint(g));
-    }
-
-    pub(super) fn exec_lcm_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let a = left.to_bigint().abs();
-        let b = right.to_bigint().abs();
-        let result = if a.is_zero() && b.is_zero() {
-            Value::Int(0)
-        } else {
-            let g = num_integer::Integer::gcd(&a, &b);
-            Value::from_bigint(&a / &g * &b)
-        };
-        self.stack.push(result);
-    }
-
-    pub(super) fn exec_infix_min_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        if matches!(&left, Value::Package(name) if name == "Any") {
-            self.stack.push(right);
-            return;
-        }
-        if matches!(&right, Value::Package(name) if name == "Any") {
-            self.stack.push(left);
-            return;
-        }
-        let left_is_failure =
-            matches!(&left, Value::Instance { class_name, .. } if class_name == "Failure");
-        let right_is_failure =
-            matches!(&right, Value::Instance { class_name, .. } if class_name == "Failure");
-        if left_is_failure {
-            self.stack.push(left);
-            return;
-        }
-        if right_is_failure {
-            self.stack.push(right);
-            return;
-        }
-        let ord = cmp_values(&left, &right);
-        self.stack.push(if ord.is_le() { left } else { right });
-    }
-
-    pub(super) fn exec_infix_max_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        if matches!(&left, Value::Package(name) if name == "Any") {
-            self.stack.push(right);
-            return;
-        }
-        if matches!(&right, Value::Package(name) if name == "Any") {
-            self.stack.push(left);
-            return;
-        }
-        let left_is_failure =
-            matches!(&left, Value::Instance { class_name, .. } if class_name == "Failure");
-        let right_is_failure =
-            matches!(&right, Value::Instance { class_name, .. } if class_name == "Failure");
-        if left_is_failure {
-            self.stack.push(left);
-            return;
-        }
-        if right_is_failure {
-            self.stack.push(right);
-            return;
-        }
-        let ord = cmp_values(&left, &right);
-        self.stack.push(if ord.is_ge() { left } else { right });
-    }
-
-    pub(super) fn exec_string_repeat_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-
-        // Warn on uninitialized type object used as repeat count
-        if let Value::Package(name) = &right
-            && name == "Int"
-            && !self.warning_suppressed()
-        {
-            self.write_warn_to_stderr(&format!(
-                "Use of uninitialized value of type {} in numeric context",
-                name
-            ));
-        }
-
-        // Whatever on RHS produces a WhateverCode closure
-        if matches!(&right, Value::Whatever) {
-            self.stack.push(self.make_x_whatevercode(left));
-            return Ok(());
-        }
-
-        let Some(n_raw) = crate::runtime::Interpreter::parse_repeat_count(&right)? else {
-            return Err(crate::runtime::Interpreter::repeat_error(
-                "X::Numeric::CannotConvert",
-                "Cannot convert Inf to Int".to_string(),
-            ));
-        };
-        let n = n_raw.max(0) as usize;
-        let src = crate::runtime::utils::coerce_to_str(&left);
-        // Guard the allocation: `str::repeat` aborts the process via
-        // `handle_alloc_error` on an absurd count (e.g. `"x" x 1e15`), which
-        // `try {}` cannot recover from. Reserve fallibly first so the same
-        // input yields a catchable `X::` instead. (raku aborts here too.)
-        let total = src
-            .len()
-            .checked_mul(n)
-            .ok_or_else(|| RuntimeError::new("Cannot repeat string: length overflow"))?;
-        let mut repeated = String::new();
-        repeated.try_reserve(total).map_err(|_| {
-            RuntimeError::new(format!(
-                "Cannot repeat string to {total} bytes: memory allocation failed"
-            ))
-        })?;
-        for _ in 0..n {
-            repeated.push_str(&src);
-        }
-        let result = if repeated.is_ascii() {
-            Value::str(repeated)
-        } else {
-            Value::str(repeated.nfc().collect::<String>())
-        };
-        self.stack.push(result);
-        Ok(())
-    }
-
-    pub(super) fn exec_list_repeat_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-
-        // Fast path for xx-reeval thunks (compiler-generated repeat blocks)
-        let thunk = match &left {
-            Value::Sub(data) => Some(data.clone()),
-            Value::WeakSub(weak) => weak.upgrade(),
-            _ => None,
-        };
-        if let Some(data) = thunk.filter(|data| Self::is_xx_reeval_thunk(data.as_ref()))
-            && let Value::Int(n) = right
-        {
-            let n = n.max(0) as usize;
-            // Fast path: @var.shift xx N or @var.pop xx N — bulk drain
-            if let Some(items) = self.try_bulk_shift_pop(&data, n)? {
-                self.stack.push(Value::Seq(Arc::new(items)));
-                return Ok(());
-            }
-            let items = self.vm_xx_repeat_thunk(data.as_ref(), n)?;
-            self.stack.push(Value::Seq(Arc::new(items)));
-            return Ok(());
-        }
-
-        // Warn on uninitialized type object used as repeat count
-        if let Value::Package(name) = &right
-            && name == "Int"
-            && !self.warning_suppressed()
-        {
-            self.write_warn_to_stderr(&format!(
-                "Use of uninitialized value of type {} in numeric context",
-                name
-            ));
-        }
-
-        const EAGER_LIMIT: usize = 10_000;
-        const LAZY_CACHE: usize = 4_096;
-        const LAZY_CACHE_CALLABLE: usize = 256;
-
-        let is_callable = matches!(
-            left,
-            Value::Sub(_) | Value::WeakSub(_) | Value::Routine { .. }
-        );
-        let lazy_cache = if is_callable {
-            LAZY_CACHE_CALLABLE
-        } else {
-            LAZY_CACHE
-        };
-
-        let count = crate::runtime::Interpreter::parse_repeat_count(&right)?;
-        let (repeat, lazy) = match count {
-            Some(n) if n <= 0 => (0usize, false),
-            Some(n) if (n as usize) <= EAGER_LIMIT => (n as usize, false),
-            Some(n) => ((n as usize).min(lazy_cache), true),
-            None => (lazy_cache, true),
-        };
-
-        let mut items = Vec::with_capacity(repeat);
-        if let Value::Slip(slip_items) = &left {
-            if slip_items.is_empty() {
-                items.extend(std::iter::repeat_n(Value::Nil, repeat));
-            } else {
-                for _ in 0..repeat {
-                    items.extend(slip_items.iter().cloned());
-                }
-            }
-        } else {
-            for _ in 0..repeat {
-                items.push(self.repeat_lhs_once(&left)?);
-            }
-        }
-
-        let result = if lazy {
-            let count = crate::runtime::Interpreter::repeat_logical_count(&right);
-            crate::runtime::Interpreter::make_repeat_lazy_cache_counted(items, count)
-        } else {
-            Value::Seq(Arc::new(items))
-        };
-        self.stack.push(result);
-        Ok(())
-    }
-
-    pub(super) fn exec_function_compose_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let composed = self.compose_callables(left, right);
-        self.stack.push(composed);
-    }
-
-    pub(super) fn exec_but_mixin_op(&mut self, code: &CompiledCode) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        // `$obj but R` runs role R's `submethod TWEAK`/`BUILD` via the interpreter,
-        // which can mutate a captured-outer caller lexical by name (`my $invoked;
-        // role R { submethod TWEAK { $invoked = True } }`). Snapshot the
-        // overwritable slots for the precise diff that reconciles them.
-        let pre_env = self.snapshot_carrier_overwritable_env(code);
-        // `but` composing a role or another type into a type-object invocant is
-        // illegal (no instance); mixing a concrete value (`Method but True`) is
-        // allowed, so this only guards the role / type-object-RHS branches.
-        let left_type_object = self.does_invocant_type_object(&left);
-        let role_composed = match &right {
-            Value::Pair(name, boxed)
-                if self.has_role(name) && matches!(boxed.as_ref(), Value::Array(..)) =>
-            {
-                Some(loan_env!(
-                    self,
-                    eval_does_values(left.clone(), right.clone())
-                ))
-            }
-            Value::Package(name) if self.has_role(&name.resolve()) => Some(loan_env!(
-                self,
-                eval_does_values(left.clone(), right.clone())
-            )),
-            Value::Str(name) if self.has_role(name) => Some(loan_env!(
-                self,
-                eval_does_values(left.clone(), right.clone())
-            )),
-            _ => None,
-        };
-        if let Some(composed) = role_composed {
-            if let Some(tn) = &left_type_object {
-                return Err(Self::does_type_object_error("but", tn));
-            }
-            let composed = composed?;
-            // Reconcile the caller's slots from env so a captured-outer write made
-            // by the role's `submethod TWEAK`/`BUILD` is visible.
-            self.carrier_writeback_changed_aggregates(code, &pre_env);
-            self.stack.push(composed);
-            return Ok(());
-        }
-        // A type object / class (not a role) on the RHS: into a type-object
-        // invocant it is X::Does::TypeObject; otherwise not composable.
-        if matches!(&right, Value::Package(_)) {
-            if let Some(tn) = &left_type_object {
-                return Err(Self::does_type_object_error("but", tn));
-            }
-            return Err(Self::mixin_not_composable_error(&left, &right));
-        }
-        let result = Self::apply_but_mixin(left, right)?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    /// Apply a `but` mixin for a single element from a tuple expansion,
-    /// with duplicate type conflict checking.
-    pub(super) fn exec_but_mixin_tuple_elem_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let mixin_type = Self::mixin_type_for_value(&right);
-        // Check for duplicate type conflict
-        if let Value::Mixin(_, ref existing) = left
-            && existing.contains_key(&mixin_type)
-        {
-            return Err(RuntimeError::new(format!(
-                "Method '{}' must be resolved by class {}+{{}} because it exists in multiple roles",
-                mixin_type,
-                crate::runtime::utils::value_type_name(&left)
-            )));
-        }
-        let result = Self::apply_single_mixin(left, mixin_type, right);
-        self.stack.push(result);
-        Ok(())
-    }
-
-    /// Determine the mixin type name for a value (used by `but` operator).
-    fn mixin_type_for_value(val: &Value) -> String {
-        match val {
-            Value::Bool(_) => "Bool".to_string(),
-            Value::Int(_) | Value::BigInt(_) => "Int".to_string(),
-            Value::Num(_) => "Num".to_string(),
-            Value::Str(_) => "Str".to_string(),
-            Value::Rat(..) | Value::FatRat(..) | Value::BigRat(..) => "Rat".to_string(),
-            Value::Complex(..) => "Complex".to_string(),
-            Value::Package(name) => name.resolve(),
-            Value::Enum { enum_type, .. } => enum_type.resolve(),
-            Value::Instance { class_name, .. } => class_name.resolve(),
-            _ => "Any".to_string(),
-        }
-    }
-
-    /// Apply a `but`-style mixin: wrap the left value with the right value
-    /// keyed by its type name.
-    fn apply_but_mixin(left: Value, right: Value) -> Result<Value, RuntimeError> {
-        // Handle Array RHS: `True but [1, 2]` generates a single "Array" mixin
-        if let Value::Array(_, kind) = &right {
-            if kind.is_real_array() {
-                return Ok(Self::apply_single_mixin(left, "Array".to_string(), right));
-            }
-            // List values from method calls like .list -> "List" mixin
-            return Ok(Self::apply_single_mixin(left, "List".to_string(), right));
-        }
-        if matches!(&right, Value::Seq(_)) {
-            return Ok(Self::apply_single_mixin(left, "List".to_string(), right));
-        }
-
-        let mixin_type = Self::mixin_type_for_value(&right);
-        Ok(Self::apply_single_mixin(left, mixin_type, right))
-    }
-
-    /// Apply a single mixin with the given type name.
-    fn apply_single_mixin(left: Value, mixin_type: String, right: Value) -> Value {
-        match left {
-            Value::Mixin(inner, existing_mixins) => {
-                let mut mixins = (*existing_mixins).clone();
-                mixins.insert(mixin_type, right);
-                Value::mixin((*inner).clone(), mixins)
-            }
-            other => {
-                let mut mixins = std::collections::HashMap::new();
-                mixins.insert(mixin_type, right);
-                Value::mixin(other, mixins)
-            }
-        }
-    }
-
-    pub(super) fn exec_isa_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let type_name = right.to_string_value();
-        let result = left.isa_check(&type_name);
-        self.stack.push(Value::Bool(result));
-    }
-
-    /// If `left` is a *built-in* type object, return its type name. Mixing a
-    /// role into such an object via `does`/`but` is illegal (X::Does::TypeObject)
-    /// because there is no instance to compose into. A *user-defined* class type
-    /// object (including an anonymous `class {}`) may have a role mixed in —
-    /// that creates a new anonymous subtype — so it is excluded here. Undefined
-    /// scalars are stored as `Nil` and act as the `Any` type object.
-    fn does_invocant_type_object(&self, left: &Value) -> Option<String> {
-        match left {
-            Value::Nil => Some("Any".to_string()),
-            Value::Package(name) => {
-                let n = name.resolve();
-                if self.has_class(&n) {
-                    None
-                } else {
-                    Some(n.to_string())
-                }
-            }
-            _ => None,
-        }
-    }
-
-    /// `does`/`but` used with a type object as the invocant is illegal —
-    /// there is no instance to mix into (`Bool does role {...}`,
-    /// `(my $x) does Int`).
-    fn does_type_object_error(op: &str, type_name: &str) -> RuntimeError {
-        let msg = format!(
-            "Cannot use '{}' operator on a type object {}.",
-            op, type_name
-        );
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("message".to_string(), Value::str(msg.clone()));
-        attrs.insert("typename".to_string(), Value::str(type_name.to_string()));
-        let mut err = RuntimeError::new(msg);
-        err.exception = Some(Box::new(Value::make_instance(
-            Symbol::intern("X::Does::TypeObject"),
-            attrs,
-        )));
-        err
-    }
-
-    /// Mixing a non-composable type (a class or type object, not a role or a
-    /// concrete value) into an object is illegal (`5 does Int`, `obj does NC`).
-    /// `target` is the object being mixed into, `rolish` the offending type.
-    fn mixin_not_composable_error(target: &Value, rolish: &Value) -> RuntimeError {
-        let mixin_type = rolish.to_string_value();
-        let into_type = crate::runtime::utils::value_type_name(target);
-        let msg = format!(
-            "Cannot mix in non-composable type {} into object of type {}",
-            mixin_type, into_type
-        );
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("message".to_string(), Value::str(msg.clone()));
-        attrs.insert("target".to_string(), target.clone());
-        attrs.insert("rolish".to_string(), rolish.clone());
-        let mut err = RuntimeError::new(msg);
-        err.exception = Some(Box::new(Value::make_instance(
-            Symbol::intern("X::Mixin::NotComposable"),
-            attrs,
-        )));
-        err
-    }
-
-    /// Interpreter-native `does` check. Inlines the pure `does_check` path and
-    /// only falls back to the interpreter for actual role composition.
-    fn vm_does_values(&mut self, left: Value, right: Value) -> Result<Value, RuntimeError> {
-        // `does`/`but` on a type-object invocant (undefined scalars are stored
-        // as Nil and act as the `Any` type object) is illegal *when composing a
-        // role or another type* — there is no instance to compose into. Mixing a
-        // *concrete value* (e.g. `Method but True`) is still allowed, so this
-        // check guards only the role / type-object-RHS branches below.
-        let left_type_object = self.does_invocant_type_object(&left);
-        // Handle array of roles: `$obj does (RoleA, RoleB)`
-        if let Value::Array(ref items, ..) = right {
-            let all_roles = items.iter().all(|item| self.is_role_application(item));
-            if all_roles && !items.is_empty() {
-                if let Some(tn) = &left_type_object {
-                    return Err(Self::does_type_object_error("does", tn));
-                }
-                return loan_env!(self, eval_does_values_list(left, items.as_ref()));
-            }
-        }
-        // Check if the RHS is a role that needs to be composed onto the value.
-        // If so, delegate to the interpreter which manages role state.
-        if self.is_role_application(&right) {
-            if let Some(tn) = &left_type_object {
-                return Err(Self::does_type_object_error("does", tn));
-            }
-            return loan_env!(self, eval_does_values(left, right));
-        }
-        // When the RHS is an enum value, `does` acts as a mixin (like `but`).
-        if matches!(&right, Value::Enum { .. }) {
-            return Self::apply_but_mixin(left, right);
-        }
-        // When the RHS is a concrete value (Str, Int, Bool, etc.), `does` acts
-        // as a mutating mixin — it overrides the corresponding type method.
-        // For example, `$o does "modded"` makes `$o.Str` return "modded".
-        if matches!(
-            &right,
-            Value::Str(_)
-                | Value::Int(_)
-                | Value::BigInt(_)
-                | Value::Num(_)
-                | Value::Bool(_)
-                | Value::Rat(..)
-        ) {
-            return Self::apply_but_mixin(left, right);
-        }
-        // When the RHS is an Instance, mix it in so that calling .$ClassName
-        // on the result returns that instance.
-        if matches!(&right, Value::Instance { .. }) {
-            return Self::apply_but_mixin(left, right);
-        }
-        // A type object / class (not a role) on the RHS: `(my $foo) does Int`
-        // mixes into a type-object invocant (X::Does::TypeObject); otherwise the
-        // type itself is not composable (`5 does Int`, `obj does SomeClass`).
-        if matches!(&right, Value::Package(_)) {
-            if let Some(tn) = &left_type_object {
-                return Err(Self::does_type_object_error("does", tn));
-            }
-            return Err(Self::mixin_not_composable_error(&left, &right));
-        }
-        // Pure check: does the value conform to the named role/type?
-        let role_name = right.to_string_value();
-        Ok(Value::Bool(left.does_check(&role_name)))
-    }
-
-    pub(super) fn exec_does_op(&mut self, code: &CompiledCode) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        // Sync Interpreter locals to interpreter env so BUILD submethods can access
-        // and modify closure variables from the enclosing scope.
-        self.sync_env_from_locals(code);
-        // Snapshot for the precise BUILD/TWEAK captured-outer writeback (see
-        // exec_does_var_op).
-        let pre_env = self.snapshot_carrier_overwritable_env(code);
-        let result = self.vm_does_values(left, right)?;
-        // Sync back: BUILD submethods may have modified closure variables, so the
-        // captured-outer writes reach the caller's slots.
-        self.carrier_writeback_changed_aggregates(code, &pre_env);
-        // Capture Mixin value for trait_mod writeback (same as DoesVar path)
-        if matches!(&result, Value::Mixin(..)) && self.trait_mod_writeback_key.is_some() {
-            self.trait_mod_writeback_value = Some(result.clone());
-        }
-        self.stack.push(result);
-        Ok(())
-    }
-
-    pub(super) fn exec_does_var_op(
-        &mut self,
-        code: &CompiledCode,
-        name_idx: u32,
-    ) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        // Sync Interpreter locals to interpreter env so BUILD submethods can access
-        // and modify closure variables from the enclosing scope.
-        self.sync_env_from_locals(code);
-        // A `submethod BUILD`/`TWEAK` run by the mixin can mutate a captured-outer
-        // caller lexical (`my $n=0; role R { submethod TWEAK { $n++ } }; $x does R`).
-        // Snapshot the overwritable slots so the precise diff reconciles them.
-        let pre_env = self.snapshot_carrier_overwritable_env(code);
-        let updated = self.vm_does_values(left, right)?;
-        // Sync back: BUILD submethods may have modified closure variables.
-        self.carrier_writeback_changed_aggregates(code, &pre_env);
-        let name = Self::const_str(code, name_idx).to_string();
-        self.env_mut().insert(name.clone(), updated.clone());
-        self.update_local_if_exists(code, &name, &updated);
-        // Capture Mixin value for trait_mod writeback: when `$r does Role`
-        // runs inside a trait_mod:<is>, the Mixin needs to propagate back
-        // to the outer scope's `&name` variable.
-        if matches!(&updated, Value::Mixin(..)) && self.trait_mod_writeback_key.is_some() {
-            self.trait_mod_writeback_value = Some(updated.clone());
-        }
-        self.stack.push(updated);
-        Ok(())
-    }
-
-    pub(super) fn exec_make_pair_op(&mut self, code: &crate::opcode::CompiledCode) {
-        let mut right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        // `key => $var`: the RHS scalar variable was tagged with `WrapVarRef`, so
-        // capture its container. The Pair's value becomes a `ContainerRef` shared
-        // with `$var`, giving Raku's write-through semantics: `$pair.value = X`
-        // updates `$var`, and `$pair.value<k> = v` writes through to `$var`'s
-        // backing Array/Hash. (See S02:1704 / roast S02-types/pair.t.)
-        if let Value::Capture { positional, named } = &right
-            && positional.is_empty()
-            && let Some(Value::Str(source_name)) = named.get("__mutsu_varref_name")
-            && let Some(inner) = named.get("__mutsu_varref_value")
-        {
-            let source_name = source_name.to_string();
-            let inner = inner.clone();
-            right = self.capture_var_cell(code, &source_name, inner);
-        }
-        // Preserve the original key type for `.key` to return the correct type.
-        // Use ValuePair for non-string keys, Pair for string keys.
-        match &left {
-            Value::Str(_) => {
-                let key = left.to_string_value();
-                self.stack.push(Value::Pair(key, Box::new(right)));
-            }
-            _ => {
-                self.stack
-                    .push(Value::ValuePair(Box::new(left), Box::new(right)));
-            }
-        }
-    }
-
-    pub(super) fn exec_bit_and_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let (l, r) = runtime::coerce_numeric(left, right);
-        let result = match (l, r) {
-            (Value::Int(a), Value::Int(b)) => Value::Int(a & b),
-            (Value::BigInt(a), Value::BigInt(b)) => Value::from_bigint(&*a & &*b),
-            (Value::BigInt(a), Value::Int(b)) => {
-                Value::from_bigint(&*a & &num_bigint::BigInt::from(b))
-            }
-            (Value::Int(a), Value::BigInt(b)) => {
-                Value::from_bigint(&num_bigint::BigInt::from(a) & &*b)
-            }
-            _ => Value::Int(0),
-        };
-        self.stack.push(result);
-    }
-
-    pub(super) fn exec_bit_or_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let (l, r) = runtime::coerce_numeric(left, right);
-        let result = match (l, r) {
-            (Value::Int(a), Value::Int(b)) => Value::Int(a | b),
-            (Value::BigInt(a), Value::BigInt(b)) => Value::from_bigint(&*a | &*b),
-            (Value::BigInt(a), Value::Int(b)) => {
-                Value::from_bigint(&*a | &num_bigint::BigInt::from(b))
-            }
-            (Value::Int(a), Value::BigInt(b)) => {
-                Value::from_bigint(&num_bigint::BigInt::from(a) | &*b)
-            }
-            _ => Value::Int(0),
-        };
-        self.stack.push(result);
-    }
-
-    pub(super) fn exec_bit_xor_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let (l, r) = runtime::coerce_numeric(left, right);
-        let result = match (l, r) {
-            (Value::Int(a), Value::Int(b)) => Value::Int(a ^ b),
-            (Value::BigInt(a), Value::BigInt(b)) => Value::from_bigint(&*a ^ &*b),
-            (Value::BigInt(a), Value::Int(b)) => {
-                Value::from_bigint(&*a ^ &num_bigint::BigInt::from(b))
-            }
-            (Value::Int(a), Value::BigInt(b)) => {
-                Value::from_bigint(&num_bigint::BigInt::from(a) ^ &*b)
-            }
-            _ => Value::Int(0),
-        };
-        self.stack.push(result);
-    }
-
-    pub(super) fn exec_bit_shift_left_op(&mut self) {
-        fn shift_left_i64(a: i64, b: i64) -> Value {
-            if b < 0 {
-                let shift = b.unsigned_abs();
-                let shifted = if shift >= i64::BITS as u64 {
-                    if a < 0 { -1 } else { 0 }
-                } else {
-                    a >> (shift as u32)
-                };
-                return Value::Int(shifted);
-            }
-            let shift = b as u64;
-            if shift >= i64::BITS as u64 {
-                return Value::from_bigint(num_bigint::BigInt::from(a) << (shift as usize));
-            }
-            // Use BigInt for the shift to avoid i64 overflow (Raku integers are arbitrary precision)
-            Value::from_bigint(num_bigint::BigInt::from(a) << (shift as usize))
-        }
-
-        fn shift_left_bigint(a: num_bigint::BigInt, b: i64) -> Value {
-            if b < 0 {
-                Value::from_bigint(a >> (b.unsigned_abs() as usize))
-            } else {
-                Value::from_bigint(a << (b as usize))
-            }
-        }
-
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let (l, r) = runtime::coerce_numeric(left, right);
-        let result = match (l, r) {
-            (Value::Int(a), Value::Int(b)) => shift_left_i64(a, b),
-            (l, r) => {
-                let a = l.to_bigint();
-                let b_big = r.to_bigint();
-                let b = b_big.to_i64().unwrap_or_else(|| {
-                    if b_big.sign() == num_bigint::Sign::Minus {
-                        i64::MIN
-                    } else {
-                        i64::MAX
-                    }
-                });
-                shift_left_bigint(a, b)
-            }
-        };
-        self.stack.push(result);
-    }
-
-    pub(super) fn exec_bit_shift_right_op(&mut self) {
-        fn shift_right_i64(a: i64, b: i64) -> Value {
-            if b < 0 {
-                let shift = b.unsigned_abs();
-                if shift >= i64::BITS as u64 {
-                    return Value::from_bigint(num_bigint::BigInt::from(a) << (shift as usize));
-                }
-                if let Some(v) = a.checked_shl(shift as u32) {
-                    Value::Int(v)
-                } else {
-                    Value::from_bigint(num_bigint::BigInt::from(a) << (shift as usize))
-                }
-            } else {
-                let shift = b as u64;
-                let shifted = if shift >= i64::BITS as u64 {
-                    if a < 0 { -1 } else { 0 }
-                } else {
-                    a >> (shift as u32)
-                };
-                Value::Int(shifted)
-            }
-        }
-
-        fn shift_right_bigint(a: num_bigint::BigInt, b: i64) -> Value {
-            if b < 0 {
-                Value::from_bigint(a << (b.unsigned_abs() as usize))
-            } else {
-                Value::from_bigint(a >> (b as usize))
-            }
-        }
-
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let (l, r) = runtime::coerce_numeric(left, right);
-        let result = match (l, r) {
-            (Value::Int(a), Value::Int(b)) => shift_right_i64(a, b),
-            (l, r) => {
-                let a = l.to_bigint();
-                let b_big = r.to_bigint();
-                let b = b_big.to_i64().unwrap_or_else(|| {
-                    if b_big.sign() == num_bigint::Sign::Minus {
-                        i64::MIN
-                    } else {
-                        i64::MAX
-                    }
-                });
-                shift_right_bigint(a, b)
-            }
-        };
-        self.stack.push(result);
-    }
-
-    pub(super) fn exec_bool_bit_or_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        self.stack.push(Value::Bool(left.truthy() | right.truthy()));
-    }
-
-    pub(super) fn exec_bool_bit_and_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        self.stack.push(Value::Bool(left.truthy() & right.truthy()));
-    }
-
-    pub(super) fn exec_bool_bit_xor_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        self.stack.push(Value::Bool(left.truthy() ^ right.truthy()));
-    }
-
-    /// String bitwise AND (~&): AND corresponding bytes of two strings.
-    pub(super) fn exec_str_bit_and_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let l = left.to_string_value();
-        let r = right.to_string_value();
-        let lb = l.as_bytes();
-        let rb = r.as_bytes();
-        let len = lb.len().max(rb.len());
-        let result: Vec<u8> = (0..len)
-            .map(|i| {
-                let a = lb.get(i).copied().unwrap_or(0);
-                let b = rb.get(i).copied().unwrap_or(0);
-                a & b
-            })
-            .collect();
-        self.stack
-            .push(Value::str(String::from_utf8_lossy(&result).into_owned()));
-    }
-
-    /// String bitwise OR (~|): OR corresponding bytes of two strings.
-    pub(super) fn exec_str_bit_or_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let l = left.to_string_value();
-        let r = right.to_string_value();
-        let lb = l.as_bytes();
-        let rb = r.as_bytes();
-        let len = lb.len().max(rb.len());
-        let result: Vec<u8> = (0..len)
-            .map(|i| {
-                let a = lb.get(i).copied().unwrap_or(0);
-                let b = rb.get(i).copied().unwrap_or(0);
-                a | b
-            })
-            .collect();
-        self.stack
-            .push(Value::str(String::from_utf8_lossy(&result).into_owned()));
-    }
-
-    /// String bitwise XOR (~^): XOR corresponding bytes of two strings.
-    pub(super) fn exec_str_bit_xor_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let l = left.to_string_value();
-        let r = right.to_string_value();
-        let lb = l.as_bytes();
-        let rb = r.as_bytes();
-        let len = lb.len().max(rb.len());
-        let result: Vec<u8> = (0..len)
-            .map(|i| {
-                let a = lb.get(i).copied().unwrap_or(0);
-                let b = rb.get(i).copied().unwrap_or(0);
-                a ^ b
-            })
-            .collect();
-        self.stack
-            .push(Value::str(String::from_utf8_lossy(&result).into_owned()));
-    }
-
-    /// String bitwise shift left (~<): not yet implemented in Raku either.
-    /// TODO: Implement when Raku spec is finalized for this operator.
-    pub(super) fn exec_str_shift_left_op(&mut self) {
-        let _right = self.stack.pop().unwrap();
-        let _left = self.stack.pop().unwrap();
-        self.stack.push(Value::Nil);
-    }
-
-    /// String bitwise shift right (~>): not yet implemented in Raku either.
-    /// TODO: Implement when Raku spec is finalized for this operator.
-    pub(super) fn exec_str_shift_right_op(&mut self) {
-        let _right = self.stack.pop().unwrap();
-        let _left = self.stack.pop().unwrap();
-        self.stack.push(Value::Nil);
     }
 }
